Move generate_system into new type
